+++ src/tallies/generic/includes/tallyAngularDetector.hh
@@ -66,26 +66,26 @@ public:
   void scapedParticle(const pen_KPAR kpar,
 				 const pen_particleState& state);
   
-  void tally_move2geo(const double /*nhist*/,
+  void tally_move2geo(const unsigned long long /*nhist*/,
 		      const unsigned /*kdet*/,
 		      const pen_KPAR kpar,
 		      const pen_particleState& state,
 		      const double /*dsef*/,
 		      const double /*dstot*/);
   
-  void tally_matChange(const double /*nhist*/,
+  void tally_matChange(const unsigned long long /*nhist*/,
 		       const pen_KPAR kpar,
 		       const pen_particleState& state,
 		       const unsigned /*prevMat*/);
   
-  void tally_endHist(const double /*nhist*/);
+  void tally_endHist(const unsigned long long /*nhist*/);
   
   
   int configure(const wrapper_geometry& /*geometry*/,
 		const abc_material* const /*materials*/[constants::MAXMAT],     
 		const pen_parserSection& config, const unsigned verbose);
   void flush();
-  void saveData(const double nhist) const;
+  void saveData(const unsigned long long nhist) const;
   int sumTally(const pen_AngularDet& tally);
   
 };
