+++ src/tallies/generic/includes/tallyEnergyDepositionMat.hh
@@ -52,38 +52,38 @@ public:
 		  nmat(0)
   {}
   
-  void tally_localEdep(const double /*nhist*/,
+  void tally_localEdep(const unsigned long long /*nhist*/,
 		       const pen_KPAR /*kpar*/,
 		       const pen_particleState& state,
 		       const double dE);
-  void tally_beginPart(const double /*nhist*/,
+  void tally_beginPart(const unsigned long long /*nhist*/,
 		       const unsigned /*kdet*/,
 		       const pen_KPAR /*kpar*/,
 		       const pen_particleState& state);
-  void tally_beginHist(const double /*nhist*/,
+  void tally_beginHist(const unsigned long long /*nhist*/,
 		       const unsigned /*kdet*/,
 		       const pen_KPAR /*kpar*/,
 		       const pen_particleState& state);
 
-  void tally_step(const double /*nhist*/,
+  void tally_step(const unsigned long long /*nhist*/,
 		  const pen_KPAR /*kpar*/,
 		  const pen_particleState& state,
 		  const tally_StepData& stepData);
   
-  void tally_move2geo(const double /*nhist*/,
+  void tally_move2geo(const unsigned long long /*nhist*/,
 		      const unsigned /*kdet*/,
 		      const pen_KPAR /*kpar*/,
 		      const pen_particleState& state,
 		      const double /*dsef*/,
 		      const double /*dstot*/);
     
-  void tally_endHist(const double /*nhist*/);
+  void tally_endHist(const unsigned long long /*nhist*/);
 
   int configure(const wrapper_geometry& /*geometry*/,
 		const abc_material* const /*materials*/[constants::MAXMAT],     
 		const pen_parserSection& config, const unsigned verbose);
   void flush();
-  void saveData(const double nhist) const;
+  void saveData(const unsigned long long nhist) const;
   int sumTally(const pen_EdepMat& tally);
 };
