+++ src/tallies/generic/includes/tallySpatialDoseDistrib.hh
@@ -40,13 +40,13 @@ class pen_SpatialDoseDistrib: public pen_genericTally<pen_particleState> {
   long int nxy,nbin;
   double xmin,ymin,zmin,dx,dy,dz,idx,idy,idz;
   double voxVol;
-  double* nlast;
+  unsigned long long* nlast;
   double* edptmp;
   double* edep;
   double* edep2;
   double* ivoxMass;
 
-  double* nlastdepth;
+  unsigned long long* nlastdepth;
   double* edepthtmp;
   double* edepth;
   double* edepth2;
@@ -81,7 +81,7 @@ public:
 }
   
   void updateEdepCounters(const double dE,
-                          const double nhist,
+                          const unsigned long long nhist,
                           const double X,
 			  const double Y,
 			  const double Z,
@@ -90,27 +90,27 @@ public:
     
   void clear(void);
   
-  void tally_localEdep(const double nhist,
+  void tally_localEdep(const unsigned long long nhist,
 		       const pen_KPAR /*kpar*/,
 		       const pen_particleState& state,
 		       const double dE);
   
-  void tally_beginPart(const double nhist,
+  void tally_beginPart(const unsigned long long nhist,
 		       const unsigned /*kdet*/,
 		       const pen_KPAR /*kpar*/,
 		       const pen_particleState& state);
   
-  void tally_beginHist(const double nhist,
+  void tally_beginHist(const unsigned long long nhist,
 		       const unsigned /*kdet*/,
 		       const pen_KPAR /*kpar*/,
 		       const pen_particleState& state);
 
-  void tally_step(const double nhist,
+  void tally_step(const unsigned long long nhist,
 		  const pen_KPAR /*kpar*/,
 		  const pen_particleState& state,
 		  const tally_StepData& stepData);
   
-  void tally_move2geo(const double nhist,
+  void tally_move2geo(const unsigned long long nhist,
 		      const unsigned /*kdet*/,
 		      const pen_KPAR /*kpar*/,
 		      const pen_particleState& state,
@@ -118,13 +118,13 @@ public:
 		      const double /*dstot*/);
   
   
-  void tally_endSim(const double /*nhist*/);
+  void tally_endSim(const unsigned long long /*nhist*/);
     
   int configure(const wrapper_geometry& geometry,
 		const abc_material* const materials[constants::MAXMAT],
 		const pen_parserSection& config,
 		const unsigned verbose);
-  void saveData(const double nhist) const;
+  void saveData(const unsigned long long nhist) const;
   void flush(void);
   int sumTally(const pen_SpatialDoseDistrib& tally);
